Enhanced expression extraction
Changed logic for extracting info about expressions from a source line. Moved the cBaseFuncLib property paExpressions to be aa new member of the tTokenizer = aExpressions. This is a struct array with the expression, length & starting position in the source line. Note that there are also the sLeftExpression1 et al. string expression members. So some info are duplicated to make refactoring function coding easier.