Added a comment above the "dummy" procedure code
When a registered function from the Functions table is found that is not found in oRefactor_Unit_Tests.pkg, a dummy object and Procedure is written to the .json file. The newly added comment on top of that procedure explains what to do.
Also skipped the Procedure name '_xx' part to to '_1'